Finished setting up basic Ember Simple Auth
New and Edit buttons only accessible if authenticated.

Because our AppicationRoute overrides beforeModel we had to add a call to _super and pass in the transition for Ember Simple Auth to be able to transiton properly based on login state, i.e. transition transition you to the login URL when you are not authenticated and transition you back to the URL you tried to visit before being authenticated.